refactor(web): group blog post list component
