+++ web/next/content/docs/manage/blog.mdx
@@ -66,7 +66,7 @@ Images for a post are committed as static assets under `web/next/public/blog/<sl
 
 - **Schema**: the `blog` collection in `web/next/source.config.ts` requires `createdAt` (and accepts optional `updatedAt`/`publishedAt`/`draft`/`author`/`tags`), so every post is validated at build time. Running `bun .github/scripts/docs.ts` adds a missing `createdAt` with the local date for new blog files and prints the value it wrote; review and commit that frontmatter. Strict builds validate without writing.
 - **Publishing**: `web/next/src/lib/blog-policy.ts` defines the shared rule for published posts: not draft and `publishedAt` is present and not future-dated. `web/next/src/lib/blog.ts` adapts that rule to Fumadocs pages for the blog routes, listing, sitemap, `llms.txt`, and OG image routes.
-- **Listing**: `/blog` renders the `<BlogPostList />` component (`web/next/src/components/blog-post-list.tsx`), which displays published posts newest-first by `publishedAt`. `web/next/content/blog/index.mdx` only holds the page intro plus `<BlogPostList />`.
+- **Listing**: `/blog` renders the `<BlogPostList />` component (`web/next/src/components/blog/post-list.tsx`), which displays published posts newest-first by `publishedAt`. `web/next/content/blog/index.mdx` only holds the page intro plus `<BlogPostList />`.
 - **Article metadata**: individual blog pages display their publish/update dates and emit article Open Graph metadata from the same frontmatter.
 - **meta.json**: `web/next/content/blog/meta.json` is generated by the docs generator (`.github/scripts/docs.ts`, `generateBlogMeta`) from posts published at generation time (index first, then newest-first by `publishedAt`) and is git-ignored. It drives the Fumadocs page-tree with no hand-maintenance, so never edit it by hand. A scheduled post can become public before it appears in page-tree navigation; `/blog` is the canonical public listing.
 - **updatedAt**: optional `updatedAt` values are validated like `createdAt` and are used as the blog page `lastModified` value in `sitemap.xml`.
