+++ web/next/src/mdx-components.tsx
@@ -1,7 +1,7 @@
 import defaultMdxComponents from "fumadocs-ui/mdx"
 import type { MDXComponents } from "mdx/types"
 
-import { BlogPostList } from "@/components/blog-post-list"
+import { BlogPostList } from "@/components/blog/post-list"
 
 export function getMDXComponents(components?: MDXComponents): MDXComponents {
   return {
